+++ app/Http/Controllers/TicketController.php
@@ -2,8 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\CustomerType;
+use App\Models\Status;
 use App\Models\Ticket;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Validator;
 
 class TicketController extends Controller
 {
@@ -28,7 +32,7 @@ class TicketController extends Controller
         privilegeLevel(self::config['privilege'], ONLY_SEE);
 
         // olah data
-        $parse  = $this->parseData(Ticket::select('*'), $request);
+        $parse  = $this->parseData(Ticket::select('tickets.*'), $request);
 
         // penguraian data
         $params = [
@@ -50,7 +54,42 @@ class TicketController extends Controller
      */
     public function create()
     {
-        //
+        // cek privilege
+        privilegeLevel(self::config['privilege'], CAN_CRUD);
+
+        $data = [
+            'name'          => 'EZ-'.str_pad(Ticket::withTrashed()->get()->count() + 1, 6, '0', STR_PAD_LEFT),
+            'service_info'  => null,
+            'status'        => null,
+            'customer_name' => null,
+            'phone'         => null,
+            'phone2'        => null,
+            'phone3'        => null,
+            'address'       => null,
+            'email'         => null,
+            'customer_type' => null,
+            'model'         => null,
+            'serial'        => null,
+            'warranty_no'   => null,
+            'purchase_date' => null,
+            'warranty_type' => null,
+        ];
+
+        $data = (object) $data;
+
+        // penguraian data
+        $params = [
+            'data'              => $data,
+            'data_additional'   => [
+                'status'        => Status::all(),
+                'customer_type' => CustomerType::all(),
+            ],
+            'type'              => 'create',
+            'title'             => 'Create '.self::config['name'],
+            'config'            => self::config
+        ];
+
+        return view(self::config['blade'].'.input', $params);
     }
 
     /**
@@ -61,7 +100,24 @@ class TicketController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        // cek privilege
+        privilegeLevel(self::config['privilege'], CAN_CRUD);
+
+        if($this->validateInput($request)) {
+            $hasil = Ticket::create($request->except(['flash-fill']));
+        }
+
+        // add created by
+        if ($hasil) {
+            $hasil->update([
+                'created_by' => Auth::user()->id,
+            ]);
+        }
+
+        // send result
+        $params = getStatus($hasil ? 'success' : 'error', 'create', self::config['name']);
+
+        return redirect(self::config['url'])->with($params);
     }
 
     /**
@@ -72,7 +128,12 @@ class TicketController extends Controller
      */
     public function show(Ticket $ticket)
     {
-        //
+        // penguraian data
+//        $params = [
+//            'data'      => $warranty,
+//        ];
+
+//        return view(self::config['blade'].'.show', $params);
     }
 
     /**
@@ -83,7 +144,24 @@ class TicketController extends Controller
      */
     public function edit(Ticket $ticket)
     {
-        //
+        // cek privilege
+        privilegeLevel(self::config['privilege'], CAN_CRUD);
+
+//        $warannty = Customer::find($id);
+
+        // penguraian data
+        $params = [
+            'data'              => $ticket,
+            'data_additional'   => [
+                'status'        => Status::all(),
+                'customer_type' => CustomerType::all(),
+            ],
+            'type'              => 'edit',
+            'title'             => 'Edit '.self::config['name'],
+            'config'            => self::config
+        ];
+
+        return view(self::config['blade'].'.input', $params);
     }
 
     /**
@@ -95,7 +173,26 @@ class TicketController extends Controller
      */
     public function update(Request $request, Ticket $ticket)
     {
-        //
+        // cek privilege
+        privilegeLevel(self::config['privilege'], CAN_CRUD);
+
+//        $customerType = CustomerType::find($id);
+
+        if ($this->validateInput($request, $ticket->id)){
+            $hasil = $ticket->fill($request->except(['flash-fill']))->save();
+        }
+
+        // add updated by
+        if ($hasil) {
+            $ticket->update([
+                'updated_by' => Auth::user()->id,
+            ]);
+        }
+
+        // send result
+        $params = getStatus($hasil ? 'success' : 'error', 'update', self::config['name']);
+
+        return redirect(self::config['url'])->with($params);
     }
 
     /**
@@ -106,7 +203,20 @@ class TicketController extends Controller
      */
     public function destroy(Ticket $ticket)
     {
-        //
+        // cek privilege
+        privilegeLevel(self::config['privilege'], CAN_CRUD);
+
+//        $customerType = CustomerType::find($id);
+
+        // update siapa yang menghapus
+        $ticket->update([
+            'deleted_by' => Auth::user()->id,
+        ]);
+
+        // send result
+        $params = getStatus($ticket->delete() ? 'success' : 'error', 'delete', self::config['name']);
+
+        return redirect(self::config['url'])->with($params);
     }
 
     public function trash(Request $request)
@@ -115,7 +225,7 @@ class TicketController extends Controller
         privilegeLevel(self::config['privilege'], ONLY_SEE);
 
         // olah data
-        $parse  = $this->parseData(Warranty::onlyTrashed()->select('warranties.*'), $request);
+        $parse  = $this->parseData(Ticket::onlyTrashed()->select('tickets.*'), $request);
 
         // penguraian data
         $params = [
@@ -136,11 +246,11 @@ class TicketController extends Controller
         privilegeLevel(self::config['privilege'], ALL_ACCESS);
 
         if ($id != null){
-            $hasil = Warranty::onlyTrashed()
+            $hasil = Ticket::onlyTrashed()
                 ->where('id', $id)
                 ->restore();
         } else {
-            $hasil = Warranty::onlyTrashed()->restore();
+            $hasil = Ticket::onlyTrashed()->restore();
         }
 
         // send result
@@ -155,11 +265,11 @@ class TicketController extends Controller
         privilegeLevel(self::config['privilege'], ALL_ACCESS);
 
         if ($id != null){
-            $hasil = Warranty::onlyTrashed()
+            $hasil = Ticket::onlyTrashed()
                 ->where('id', $id)
                 ->forceDelete();
         } else {
-            $hasil = Warranty::onlyTrashed()->forceDelete();
+            $hasil = Ticket::onlyTrashed()->forceDelete();
         }
 
         // send result
@@ -170,22 +280,29 @@ class TicketController extends Controller
 
     public function parseData($data, $request)
     {
+        // join
+        $data = $data->leftJoin('states', 'tickets.status', '=', 'states.id');
+
         $search = $request->search;
         if (strlen($search) > 1) {
-            $data = $data->where('name','LIKE','%'.$search.'%')
-                ->orWhere('service_info', 'LIKE', '%'.$search.'%')
-                ->orWhere('customer_name', 'LIKE', '%'.$search.'%')
-                ->orWhere('phone', 'LIKE', '%'.$search.'%')
-                ->orWhere('phone2', 'LIKE', '%'.$search.'%')
-                ->orWhere('phone3', 'LIKE', '%'.$search.'%')
-                ->orWhere('address', 'LIKE', '%'.$search.'%')
-                ->orWhere('email', 'LIKE', '%'.$search.'%')
-                ->orWhere('customer_type', 'LIKE', '%'.$search.'%')
-                ->orWhere('model', 'LIKE', '%'.$search.'%')
-                ->orWhere('serial', 'LIKE', '%'.$search.'%')
-                ->orWhere('warranty_no', 'LIKE', '%'.$search.'%')
-                ->orWhere('purchase_date', 'LIKE', '%'.$search.'%')
-                ->orWhere('warranty_type', 'LIKE', '%'.$search.'%');
+            $data = $data->where('tickets.name','LIKE','%'.$search.'%')
+                ->orWhere('tickets.service_info', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.customer_name', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.phone', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.phone2', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.phone3', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.address', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.email', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.customer_type', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.model', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.serial', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.warranty_no', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.purchase_date', 'LIKE', '%'.$search.'%')
+                ->orWhere('tickets.warranty_type', 'LIKE', '%'.$search.'%')
+                ->orWhereHas('states', function ($q) use ($search) {
+                    $q->where('states.name','LIKE','%'.$search.'%')
+                        ->orWhere('states.color', 'LIKE', '%'.$search.'%');
+                });
         }
 
         $perPage = $request->perPage ?: self::perPage;
@@ -222,18 +339,27 @@ class TicketController extends Controller
     {
         // validasi
         $rules = [
-            'model'             => 'required|min:3|max:100',
-            'serial'            => 'required|min:3|max:100|unique:warranties,serial,'.$id,
+            'service_info'              => 'required|min:3',
+            'customer_name'             => 'required|min:3|max:100',
+            'phone'                     => 'required|min:3|max:100',
+            'address'                   => 'required|min:3',
+            'model'                     => 'required|min:3|max:100',
         ];
 
         $messages = [
-            'model.required'    => 'Model wajib diisi',
-            'model.min'         => 'Model minimal 3 karakter',
-            'model.max'         => 'Model maksimal 100 karakter',
-            'serial.required'   => 'Nomor seri wajib diisi',
-            'serial.min'        => 'Nomor seri minimal 3 karakter',
-            'serial.max'        => 'Nomor seri maksimal 100 karakter',
-            'serial.unique'     => 'Nomor seri sudah terpakai',
+            'service_info.required'     => 'Keterangan servis wajib diisi',
+            'service_info.min'          => 'Keterangan servis minimal 3 karakter',
+            'customer_name.required'    => 'Nama konsumen wajib diisi',
+            'customer_name.min'         => 'Nama konsumen minimal 3 karakter',
+            'customer_name.max'         => 'Nama konsumen maksimal 100 karakter',
+            'phone.required'            => 'Nomor telp wajib diisi',
+            'phone.min'                 => 'Nomor telp minimal 3 karakter',
+            'phone.max'                 => 'Nomor telp maksimal 100 karakter',
+            'address.required'          => 'Alamat wajib diisi',
+            'address.min'               => 'Alamat minimal 3 karakter',
+            'model.required'            => 'Model wajib diisi',
+            'model.min'                 => 'Model minimal 3 karakter',
+            'model.max'                 => 'Model maksimal 100 karakter',
         ];
 
         $validator = Validator::make($request->all(), $rules, $messages);